version/jme_ai_1.6/jme_ai/requirements.txt
Flask==2.1.1
requests==2.27.1
python-dotenv==0.19.2
mysql-connector-python==8.0.27
psycopg2-binary==2.9.3
logging==0.5.1.2
